+++ workers/unity/Packages/io.improbable.gdk.core/Subscriptions/Reader.cs
@@ -17,7 +17,9 @@ namespace Improbable.Gdk.Subscriptions
         protected readonly EntityManager EntityManager;
         protected readonly Entity Entity;
         protected readonly EntityId EntityId;
+
         private static readonly uint ComponentId = ComponentDatabase.GetComponentId<TComponent>();
+        private static readonly ComponentType ComponentAuthType = ComponentDatabase.GetMetaclass<TComponent>().Authority;
 
         private Dictionary<Action<Authority>, ulong> authorityCallbackToCallbackKey;
         private Dictionary<Action<TUpdate>, ulong> updateCallbackToCallbackKey;
@@ -35,16 +37,16 @@ namespace Improbable.Gdk.Subscriptions
             }
         }
 
-        public Authority Authority
+        public bool HasAuthority
         {
             get
             {
                 if (!IsValid)
                 {
-                    throw new InvalidOperationException("Cannot read authority when Reader is not valid");
+                    throw new InvalidOperationException($"Cannot read {nameof(HasAuthority)} when Reader is not valid");
                 }
 
-                return ComponentUpdateSystem.GetAuthority(EntityId, ComponentId);
+                return EntityManager.HasComponent(Entity, ComponentAuthType);
             }
         }