+++ workers/unity/Packages/io.improbable.gdk.core/Worker/Worker.cs
@@ -40,13 +40,16 @@ namespace Improbable.Gdk.Core
         /// </summary>
         public bool IsConnected => ConnectionHandler.IsConnected();
 
+        public IReadOnlyDictionary<string, string> WorkerFlags => workerFlags;
+
         protected IConnectionHandler ConnectionHandler;
 
         // todo replace internal with real apis
-        internal readonly View View;
         internal ViewDiff ViewDiff;
 
         internal MessagesToSend MessagesToSend;
+
+        private readonly Dictionary<string, string> workerFlags = new Dictionary<string, string>();
 
         protected Worker(IConnectionHandler connectionHandler, string workerType, ILogDispatcher logDispatcher)
         {
@@ -60,7 +63,6 @@ namespace Improbable.Gdk.Core
             logDispatcher.WorkerType = workerType;
 
             MessagesToSend = connectionHandler.GetMessagesToSendContainer();
-            View = new View();
         }
 
         /// <summary>
@@ -87,6 +89,12 @@ namespace Improbable.Gdk.Core
         public void Tick()
         {
             ConnectionHandler.GetMessagesReceived(ref ViewDiff);
+
+            // Update worker flags
+            foreach (var (key, value) in ViewDiff.GetWorkerFlagChanges())
+            {
+                workerFlags[key] = value;
+            }
         }
 
         public void EnsureMessagesFlushed(NetFrameStats frameStats)
@@ -118,7 +126,7 @@ namespace Improbable.Gdk.Core
         /// <returns>The value of the flag, if it exists, null otherwise.</returns>
         public string GetWorkerFlag(string key)
         {
-            return View.GetWorkerFlag(key);
+            return workerFlags.TryGetValue(key, out var value) ? value : null;
         }
 
         public virtual void Dispose()